requirements.txt
flask
pandas
openpyxl
python-socketio[client]
requests
